+++ runtime/src/index.ts
@@ -1,7 +1,9 @@
-export { createClient, ClientOptions } from "./client/createClient";
-export { FieldsSelection } from "./client/typeSelection";
-export { NoExtraProperties } from "./types";
-export { generateGraphqlOperation, GraphqlOperation } from "./client/generateGraphqlOperation";
+export { createClient } from "./client/createClient";
+export type { ClientOptions } from "./client/createClient";
+export type { FieldsSelection } from "./client/typeSelection";
+export type { NoExtraProperties } from "./types";
+export { generateGraphqlOperation } from "./client/generateGraphqlOperation";
+export type { GraphqlOperation } from "./client/generateGraphqlOperation";
 export { linkTypeMap } from "./client/linkTypeMap";
 export { Observable } from "zen-observable-ts";
 export { createFetcher } from "./fetcher";